### 2023-06-20 9b564c0
Normalise map projection coordinates
Change internal coordinate system for orthographic & azimuthal projections to use numbers between 0 & 1, rather than km distances (with factor of pi for azimuthal). This is mainly an internal change for simplicity, but does also make setting limits easier too.
