### 2023-06-20 e702d7c
Make map wireframe plotting optional
