+++ planetmapper/common.py
@@ -1,4 +1,4 @@
-__version__ = '1.7.4'
+__version__ = '1.7.5'
 __author__ = 'Oliver King'
 __url__ = 'https://github.com/ortk95/planetmapper'
 __license__ = 'MIT'
