### 2023-06-28 f6bba24
Add script to debug crashes
For use in debugging #258
