### 2023-07-07 69288ea
Use **kwargs for common WCS arguments
