+++ planetmapper/observation.py
@@ -350,7 +350,6 @@ class Observation(BodyXY):
         s = self.angular_dist(b1, b2, c1, c2)
         arcsec_per_px = s * 60 * 60  # s = degrees/px
         r0 = self.target_diameter_arcsec / (2 * arcsec_per_px)
-        x0, y0 = wcs.world_to_pixel_values(self.target_ra, self.target_dec)
 
         return x0, y0, r0, rotation
 
@@ -386,9 +385,7 @@ class Observation(BodyXY):
         self.set_rotation(rotation)
         self.set_disc_method('wcs')
 
-    def position_from_wcs(
-        self, suppress_warnings: bool = False, validate: bool = True
-    ) -> None:
+    def position_from_wcs(self, **kwargs) -> None:
         """
         Set disc position `(x0, y0)` using WCS information in the observation's FITS
         header.
@@ -396,47 +393,33 @@ class Observation(BodyXY):
         See also :func:`disc_from_wcs`.
 
         Args:
-            suppress_warnings: Hide warnings produced by astropy when calculating WCS
-                conversions.
-            validate: Run checks to ensure the WCS conversion has appropriate RA/Dec
-                coordinate dimensions.
+            **kwargs: See :func:`disc_from_wcs` for additional arguments.
         Raises:
             ValueError: if no WCS information is found in the FITS header, or validation
                 fails.
         """
-        x0, y0, r0, rotation = self._get_disc_params_from_wcs(
-            suppress_warnings, validate
-        )
+        x0, y0, r0, rotation = self._get_disc_params_from_wcs(**kwargs)
         self.set_x0(x0)
         self.set_y0(y0)
         self.set_disc_method('wcs_position')
 
-    def rotation_from_wcs(
-        self, suppress_warnings: bool = False, validate: bool = True
-    ) -> None:
+    def rotation_from_wcs(self, **kwargs) -> None:
         """
         Set disc rotation using WCS information in the observation's FITS header.
 
         See also :func:`disc_from_wcs`.
 
         Args:
-            suppress_warnings: Hide warnings produced by astropy when calculating WCS
-                conversions.
-            validate: Run checks to ensure the WCS conversion has appropriate RA/Dec
-                coordinate dimensions.
+            **kwargs: See :func:`disc_from_wcs` for additional arguments.
         Raises:
             ValueError: if no WCS information is found in the FITS header, or validation
                 fails.
         """
-        x0, y0, r0, rotation = self._get_disc_params_from_wcs(
-            suppress_warnings, validate
-        )
+        x0, y0, r0, rotation = self._get_disc_params_from_wcs(**kwargs)
         self.set_rotation(rotation)
         self.set_disc_method('wcs_rotation')
 
-    def plate_scale_from_wcs(
-        self, suppress_warnings: bool = False, validate: bool = True
-    ) -> None:
+    def plate_scale_from_wcs(self, **kwargs) -> None:
         """
         Set plate scale (i.e. `r0`) using WCS information in the observation's FITS
         header.
@@ -444,17 +427,12 @@ class Observation(BodyXY):
         See also :func:`disc_from_wcs`.
 
         Args:
-            suppress_warnings: Hide warnings produced by astropy when calculating WCS
-                conversions.
-            validate: Run checks to ensure the WCS conversion has appropriate RA/Dec
-                coordinate dimensions.
+            **kwargs: See :func:`disc_from_wcs` for additional arguments.
         Raises:
             ValueError: if no WCS information is found in the FITS header, or validation
                 fails.
         """
-        x0, y0, r0, rotation = self._get_disc_params_from_wcs(
-            suppress_warnings, validate
-        )
[…]
