### 2023-07-07 ec93755
Add *args option for WCS methods
