### 2023-07-07 70f3f66
Make GUI lambdas more explicit
