### 2023-07-07 829ab86
Add header offset functionality
