### 2023-07-07 504a1c8
Adding tests for header offset code
