### 2023-07-10 7672925
Fix debugging for python 3.10
