### 2023-07-10 5718e3d
Make keep open checkbox functional in GUI
