### 2023-07-10 1c68747
Replace `plt.figure` with `Figure` for GUI plots
Using `Figure` to produce figures avoids possible segmentation fault after calling `plt.close(fig)` when a tkinter GUI is open.

Fixes #258
