### 2023-07-10 775643b
Remove redundant close_window calls
Already handled in SavingProgress
