### 2023-07-11 #5569 @ortk95
Hi! From the [list of reviewers](https://reviewers.joss.theoj.org/reviewers/search?language=26&area_id=1&keywords=planetary&name=&commit=Update+search) is seems like steo85it, jrenaud90, athulpg007 & TomGoffrey could be good fits.
