### 2023-07-20 a460eb3
Refactor projection code to simplify adding limits
