### 2023-07-21 9e4228d
Improve example in documentation
