### 2023-07-24 03526a2
Add azimuthal equal area projection shortcut
#223
