### 2023-07-24 a646ed2
Document azimuthal equal area proejction
