+++ planetmapper/body_xy.py
@@ -1751,6 +1751,9 @@ class BodyXY(Body):
         - `'azimuthal'`: azimuthal equidistant projection where the central longitude
           and latitude can be customized with the `lon` and `lat` arguments. The size of
           the map can be controlled with the `size` argument.
+        - `'azimuthal equal area'`: Lambert azimuthal equal area projection where the
+          central longitude and latitude can be customized with the `lon` and `lat` 
+          arguments. The size of the map can be controlled with the `size` argument.
         - `'manual'`: manually specify the longitude and latitude coordinates to use
           for each point on the map with the `lon_coords` and `lat_coords` arguments.
 
@@ -1792,10 +1795,12 @@ class BodyXY(Body):
             projection: String describing map projection to use (see list of supported
                 projections above).
             degree_interval: Degree interval for `'rectangular` projection.
-            lon: Central longitude of `'orthographic'` and `'azimuthal'` projections.
-            lat: Central latitude of `'orthographic'` and `'azimuthal'` projections.
-            size: Pixel size (width and height) of generated `'orthographic'` and
-                `'azimuthal'` projections.
+            lon: Central longitude of `'orthographic'`, `'azimuthal'` and `'azimuthal 
+                equal area'` projections.
+            lat: Central latitude of `'orthographic'`, `'azimuthal'` and `'azimuthal 
+                equal area'` projections.
+            size: Pixel size (width and height) of generated `'orthographic'`, 
+                `'azimuthal'` and `'azimuthal equal area'` projections.
             lon_coords: Longitude coordinates to use for `'manual'` projection. This
                 must be a tuple (e.g. use `lon_coords=tuple(np.linspace(0, 360, 100))`)
                 - this allows mapping arguments and outputs to be cached).
@@ -1903,8 +1908,8 @@ class BodyXY(Body):
                 proj, np.linspace(-lim, lim, size)
             )
             info = dict(projection=projection, lon=lon, lat=lat, size=size)
-            # XXX document
             # XXX add tests
+            # XXX add GUI option
         else:
             if projection_x_coords is None:
                 raise ValueError('x coords must be provided')
