+++ tests/test_body_xy.py
@@ -5,13 +5,14 @@ import common_testing
 import matplotlib.pyplot as plt
 import numpy as np
 from matplotlib.collections import QuadMesh
-from numpy import array, nan
+from numpy import array, inf, nan
 
 import planetmapper
 import planetmapper.base
 import planetmapper.progress
 from planetmapper import BodyXY
 from planetmapper.body_xy import Backplane, BackplaneNotFoundError
+from planetmapper.body_xy import _MapKwargs as MapKwargs
 
 
 class TestFunctions(unittest.TestCase):
@@ -582,6 +583,19 @@ class TestBodyXY(unittest.TestCase):
         fig, ax = plt.subplots()
         self.body.plot_map_wireframe(
             ax=ax,
+            projection='azimuthal equal area',
+            lat=-90,
+            label_poles=False,
+            grid_interval=45,
+        )
+        self.assertEqual(len(ax.get_lines()), 20)
+        self.assertEqual(len(ax.get_images()), 0)
+        self.assertEqual(len(ax.get_children()), 30)
+        plt.close(fig)
+
+        fig, ax = plt.subplots()
+        self.body.plot_map_wireframe(
+            ax=ax,
             projection='manual',
             lon_coords=np.linspace(-180, 180, 5),
             lat_coords=np.linspace(0, 90, 3),
@@ -859,6 +873,193 @@ class TestBodyXY(unittest.TestCase):
                 self.assertTrue(np.array_equal(yy, lats_expected))
                 self.assertEqual(info['xlim'], xlim)
                 self.assertEqual(info['ylim'], ylim)
+
+        # Test reference
+        args: list[tuple[MapKwargs, np.ndarray, np.ndarray, np.ndarray, np.ndarray]] = [
+            (
+                MapKwargs(degree_interval=123),
+                array([[307.5, 184.5, 61.5]]),
+                array([[-28.5, -28.5, -28.5]]),
+                array([[307.5, 184.5, 61.5]]),
+                array([[-28.5, -28.5, -28.5]]),
+            ),
+            (
+                MapKwargs(projection='orthographic', size=3),
+                array([[inf, inf, inf], [inf, 0.0, inf], [inf, inf, inf]]),
+                array([[inf, inf, inf], [inf, 0.0, inf], [inf, inf, inf]]),
+                array([[-1.01, 0.0, 1.01], [-1.01, 0.0, 1.01], [-1.01, 0.0, 1.01]]),
+                array([[-1.01, -1.01, -1.01], [0.0, 0.0, 0.0], [1.01, 1.01, 1.01]]),
+            ),
+            (
+                MapKwargs(projection='orthographic', size=3, lon=123.456, lat=-2),
+                array([[inf, inf, inf], [inf, 123.456, inf], [inf, inf, inf]]),
+                array([[inf, inf, inf], [inf, -2.29643357, inf], [inf, inf, inf]]),
+                array([[-1.01, 0.0, 1.01], [-1.01, 0.0, 1.01], [-1.01, 0.0, 1.01]]),
+                array([[-1.01, -1.01, -1.01], [0.0, 0.0, 0.0], [1.01, 1.01, 1.01]]),
+            ),
+            (
+                MapKwargs(projection='azimuthal', size=4),
+                array(
+                    [
+                        [inf, inf, inf, inf],
+                        [inf, -83.93213465, 83.93213465, inf],
+                        [inf, -83.93213465, 83.93213465, inf],
+                        [inf, inf, inf, inf],
+                    ]
+                ),
+                array(
+                    [
+                        [inf, inf, inf, inf],
+                        [inf, -44.83904649, -44.83904649, inf],
+                        [inf, 44.83904649, 44.83904649, inf],
[…]
