### 2023-07-24 f02c259
Add local solar time functions to `Body`
