### 2023-07-24 ed169cb
Improve documentation to work with sphinx
