+++ planetmapper/progress.py
@@ -157,6 +157,7 @@ class _SaveNavProgressHook(_SaveProgressHook):
             'BodyXY._get_radec_img': 5,
             'BodyXY._get_lonlat_centric_img': 5,
             'BodyXY._get_illumination_gie_img': 50,
+            'BodyXY.get_local_solar_time_img': 5,
             'BodyXY._get_state_imgs': 30,
             'BodyXY.get_radial_velocity_img': 5,
             'BodyXY._get_ring_coordinate_imgs': 50,
@@ -180,6 +181,7 @@ class _SaveMapProgressHook(_SaveProgressHook):
             'BodyXY._get_radec_map': 10,
             'BodyXY._get_xy_map': 10,
             # 'BodyXY._get_lonlat_map': 10,
+            'BodyXY.get_local_solar_time_map': 5,
             'BodyXY._get_lonlat_centric_map': 5,
             'BodyXY._get_state_maps': 5,
             'BodyXY._get_ring_coordinate_maps': 10,
