### 2023-07-25 05e625d
Add limb coordinate calculation
Initial work for #19
