### 2023-07-25 96eb12c
Document self.radii
