### 2023-07-25 c43cace
Cache obsvec intermediate backplanes
