### 2023-07-25 4a1468c
Add progress benchmarks
