+++ planetmapper/progress.py
@@ -160,6 +160,7 @@ class _SaveNavProgressHook(_SaveProgressHook):
             'BodyXY.get_local_solar_time_img': 5,
             'BodyXY._get_state_imgs': 30,
             'BodyXY.get_radial_velocity_img': 5,
+            'BodyXY._get_limb_coordinate_imgs': 20,
             'BodyXY._get_ring_coordinate_imgs': 50,
             'Observation.save_observation': 100,
         }
@@ -184,6 +185,7 @@ class _SaveMapProgressHook(_SaveProgressHook):
             'BodyXY.get_local_solar_time_map': 5,
             'BodyXY._get_lonlat_centric_map': 5,
             'BodyXY._get_state_maps': 5,
+            'BodyXY._get_limb_coordinate_maps': 10,
             'BodyXY._get_ring_coordinate_maps': 10,
             'Observation.save_mapped_observation': 20,
         }
