### 2023-07-25 c1b45f7
Update docstrings to more generic mapping
