+++ planetmapper/body_xy.py
@@ -76,12 +76,10 @@ class Backplane(NamedTuple):
         get_img: Function which takes no arguments returns a numpy array containing a
             backplane image when called. This should generally be a method such as
             :func:`BodyXY.get_lon_img`.
-        get_map: Function returns a numpy array containing a cylindrical map of
-            backplane values when called. This should take a single `float` argument,
-            `degree_interval` which defines the interval in degrees between the
-            longitude/latitude points in the mapped output. `degree_interval` should be
-            optional with a default value of 1. This function should generally be a
-            method such as :func:`BodyXY.get_lon_map`.
+        get_map: Function returns a numpy array containing a map of backplane values
+            when called. This should take map projection keyword arguments, as described
+            in :func:`BodyXY.generate_map_coordinates`. This function should generally
+            be a method such as :func:`BodyXY.get_lon_map`.
     """
 
     name: str
@@ -2121,7 +2119,7 @@ class BodyXY(Body):
         :func:`get_backplane_map`.
 
         Returns:
-            Array containing cylindrical map of planetographic longitude values.
+            Array containing map of planetographic longitude values.
         """
         return self._get_lonlat_map(**map_kwargs)[:, :, 0]
 
@@ -2141,7 +2139,7 @@ class BodyXY(Body):
         :func:`get_backplane_map`.
 
         Returns:
-            Array containing cylindrical map of planetographic latitude values.
+            Array containing map of planetographic latitude values.
         """
         return self._get_lonlat_map(**map_kwargs)[:, :, 1]
 
@@ -2177,7 +2175,7 @@ class BodyXY(Body):
         :func:`get_backplane_map`.
 
         Returns:
-            Array containing cylindrical map of planetocentric longitude values.
+            Array containing map of planetocentric longitude values.
         """
         return self._get_lonlat_centric_map(**map_kwargs)[:, :, 0]
 
@@ -2197,7 +2195,7 @@ class BodyXY(Body):
         :func:`get_backplane_map`.
 
         Returns:
-            Array containing cylindrical map of planetocentric latitude values.
+            Array containing map of planetocentric latitude values.
         """
         return self._get_lonlat_centric_map(**map_kwargs)[:, :, 1]
 
@@ -2235,8 +2233,8 @@ class BodyXY(Body):
         :func:`get_backplane_map`.
 
         Returns:
-            Array containing cylindrical map of right ascension values as viewed by the
-            observer. Locations which are not visible have a value of NaN.
+            Array containing map of right ascension values as viewed by the observer.
+            Locations which are not visible have a value of NaN.
         """
         return self._get_radec_map(**map_kwargs)[:, :, 0]
 
@@ -2255,8 +2253,8 @@ class BodyXY(Body):
         :func:`get_backplane_map`.
 
         Returns:
-            Array containing cylindrical map of declination values as viewed by the
-            observer. Locations which are not visible have a value of NaN.
+            Array containing map of declination values as viewed by the observer.
+            Locations which are not visible have a value of NaN.
         """
         return self._get_radec_map(**map_kwargs)[:, :, 1]
 
@@ -2291,9 +2289,9 @@ class BodyXY(Body):
         :func:`get_backplane_map`.
 
         Returns:
-            Array containing cylindrical map of the x pixel coordinates each location
-            corresponds to in the observation. Locations which are not visible or are
-            not in the image frame have a value of NaN.
+            Array containing map of the x pixel coordinates each location corresponds to
+            in the observation. Locations which are not visible or are not in the image
+            frame have a value of NaN.
         """
         return self._get_xy_map(**map_kwargs)[:, :, 0]
 
@@ -2315,9 +2313,9 @@ class BodyXY(Body):
[…]
