### 2023-07-25 c8b60fb
Correct obsvec img backplane function
