+++ tests/test_body_xy.py
@@ -914,9 +914,13 @@ class TestBodyXY(unittest.TestCase):
             'INCIDENCE: Incidence angle [deg]',
             'EMISSION: Emission angle [deg]',
             'AZIMUTH: Azimuth angle [deg]',
+            'LOCAL-SOLAR-TIME: Local solar time [local hours]',
             'DISTANCE: Distance to observer [km]',
             'RADIAL-VELOCITY: Radial velocity away from observer [km/s]',
             'DOPPLER: Doppler factor, sqrt((1 + v/c)/(1 - v/c)) where v is radial velocity',
+            'LIMB-DISTANCE: Distance above limb [km]',
+            'LIMB-LON-GRAPHIC: Planetographic longitude of closest point on the limb [deg]',
+            'LIMB-LAT-GRAPHIC: Planetographic latitude of closest point on the limb [deg]',
             'RING-RADIUS: Equatorial (ring) plane radius [km]',
             'RING-LON-GRAPHIC: Equatorial (ring) plane planetographic longitude [deg]',
             'RING-DISTANCE: Equatorial (ring) plane distance to observer [km]',
