### 2023-07-25 024a9a4
Fix out-of-bounds mapping backplanes
