### 2023-07-25 e6cae2b
Update test ref. files for newly added backplanes
