### 2023-07-25 f4f5080
Increase tolerance of unit tests
Tests were failing on GitHub Actions due to floating point rounding errors
