### 2023-07-25 6dff2e6
Remove redundant check
