### 2023-07-26 fc65e2b
Add tests for NaN propagation
For #266 - now just need to implement the NaN propagation to make the tests pass...
