+++ tests/test_body.py
@@ -380,10 +380,16 @@ class TestBody(unittest.TestCase):
             [(0, 90), (196.37390490466322, -5.561534444253404)],
             [(0, 0), (196.36982789576643, -5.565060944053696)],
             [(123.456, -56.789), (196.3691609381441, -5.5685956879058764)],
+            [(np.nan, np.nan), (np.nan, np.nan)],
+            [(np.nan, 0), (np.nan, np.nan)],
+            [(0, np.nan), (np.nan, np.nan)],
+            [(np.inf, np.inf), (np.nan, np.nan)],
         ]
         for lonlat, radec in pairs:
             with self.subTest(lonlat):
-                self.assertTrue(np.allclose(self.body.lonlat2radec(*lonlat), radec))
+                self.assertTrue(
+                    np.allclose(self.body.lonlat2radec(*lonlat), radec, equal_nan=True)
+                )
 
     def test_radec2lonlat(self):
         self.assertTrue(
@@ -404,6 +410,10 @@ class TestBody(unittest.TestCase):
                 (196.3742715121965, -5.561743939677709),
                 (180.00086055026196, 80.00042229835671),
             ],
+            [(np.nan, np.nan), (np.nan, np.nan)],
+            [(np.nan, 0), (np.nan, np.nan)],
+            [(0, np.nan), (np.nan, np.nan)],
+            [(np.inf, np.inf), (np.nan, np.nan)],
         ]
         for radec, lonlat in pairs:
             with self.subTest(radec):
@@ -411,14 +421,15 @@ class TestBody(unittest.TestCase):
                     np.allclose(
                         self.body.radec2lonlat(*radec),
                         lonlat,
+                        equal_nan=True,
                     )
                 )
-                self.assertTrue(
-                    np.allclose(
-                        self.body.lonlat2radec(*lonlat),
-                        radec,
+                if all(np.isfinite(x) for x in radec):
+                    self.assertTrue(
+                        np.allclose(
+                            self.body.lonlat2radec(*lonlat), radec, equal_nan=True
+                        )
                     )
-                )
 
     def test_lonlat2targvec(self):
         pairs = [
@@ -429,11 +440,17 @@ class TestBody(unittest.TestCase):
                 (-80, -12.3456789),
                 np.array([[12162.32647743, 68975.98103572, -13405.21131042]]),
             ),
+            [(np.nan, np.nan), np.array([np.nan, np.nan, np.nan])],
+            [(np.nan, 0), np.array([np.nan, np.nan, np.nan])],
+            [(0, np.nan), np.array([np.nan, np.nan, np.nan])],
+            [(np.inf, np.inf), np.array([np.nan, np.nan, np.nan])],
         ]
         for (lon, lat), targvec in pairs:
             with self.subTest((lon, lat)):
                 self.assertTrue(
-                    np.allclose(self.body.lonlat2targvec(lon, lat), targvec)
+                    np.allclose(
+                        self.body.lonlat2targvec(lon, lat), targvec, equal_nan=True
+                    )
                 )
 
     def test_targvec2lonlat(self):
@@ -441,10 +458,19 @@ class TestBody(unittest.TestCase):
             (np.array([0, 0, 0]), (0.0, 90.0)),
             (np.array([1, 2, 3]), (296.565051177078, 89.98665551067639)),
             (np.array([-9876, 543210, 0]), (268.9584308375042, 0.0)),
+            (np.array([np.nan, np.nan, np.nan]), (np.nan, np.nan)),
+            (np.array([np.nan, 0, 0]), (np.nan, np.nan)),
+            (np.array([0, np.nan, 0]), (np.nan, np.nan)),
+            (np.array([0, 0, np.nan]), (np.nan, np.nan)),
[…]
