### 2023-07-26 f4ba898
Standardise invalid map coords to NaN
