+++ tests/test_body_xy.py
@@ -900,15 +900,15 @@ class TestBodyXY(unittest.TestCase):
             ),
             (
                 MapKwargs(projection='orthographic', size=3),
-                array([[inf, inf, inf], [inf, 0.0, inf], [inf, inf, inf]]),
-                array([[inf, inf, inf], [inf, 0.0, inf], [inf, inf, inf]]),
+                array([[nan, nan, nan], [nan, 0.0, nan], [nan, nan, nan]]),
+                array([[nan, nan, nan], [nan, 0.0, nan], [nan, nan, nan]]),
                 array([[-1.01, 0.0, 1.01], [-1.01, 0.0, 1.01], [-1.01, 0.0, 1.01]]),
                 array([[-1.01, -1.01, -1.01], [0.0, 0.0, 0.0], [1.01, 1.01, 1.01]]),
             ),
             (
                 MapKwargs(projection='orthographic', size=3, lon=123.456, lat=-2),
-                array([[inf, inf, inf], [inf, 123.456, inf], [inf, inf, inf]]),
-                array([[inf, inf, inf], [inf, -2.29643357, inf], [inf, inf, inf]]),
+                array([[nan, nan, nan], [nan, 123.456, nan], [nan, nan, nan]]),
+                array([[nan, nan, nan], [nan, -2.29643357, nan], [nan, nan, nan]]),
                 array([[-1.01, 0.0, 1.01], [-1.01, 0.0, 1.01], [-1.01, 0.0, 1.01]]),
                 array([[-1.01, -1.01, -1.01], [0.0, 0.0, 0.0], [1.01, 1.01, 1.01]]),
             ),
@@ -916,18 +916,18 @@ class TestBodyXY(unittest.TestCase):
                 MapKwargs(projection='azimuthal', size=4),
                 array(
                     [
-                        [inf, inf, inf, inf],
-                        [inf, -83.93213465, 83.93213465, inf],
-                        [inf, -83.93213465, 83.93213465, inf],
-                        [inf, inf, inf, inf],
+                        [nan, nan, nan, nan],
+                        [nan, -83.93213465, 83.93213465, nan],
+                        [nan, -83.93213465, 83.93213465, nan],
+                        [nan, nan, nan, nan],
                     ]
                 ),
                 array(
                     [
-                        [inf, inf, inf, inf],
-                        [inf, -44.83904649, -44.83904649, inf],
-                        [inf, 44.83904649, 44.83904649, inf],
-                        [inf, inf, inf, inf],
+                        [nan, nan, nan, nan],
+                        [nan, -44.83904649, -44.83904649, nan],
+                        [nan, 44.83904649, 44.83904649, nan],
+                        [nan, nan, nan, nan],
                     ]
                 ),
                 array(
@@ -951,18 +951,18 @@ class TestBodyXY(unittest.TestCase):
                 MapKwargs(projection='azimuthal', size=4, lat=90, lon=123.456),
                 array(
                     [
-                        [inf, inf, inf, inf],
-                        [inf, 78.456, 168.456, inf],
-                        [inf, -11.544, -101.544, inf],
-                        [inf, inf, inf, inf],
+                        [nan, nan, nan, nan],
+                        [nan, 78.456, 168.456, nan],
+                        [nan, -11.544, -101.544, nan],
+                        [nan, nan, nan, nan],
                     ]
                 ),
                 array(
                     [
-                        [inf, inf, inf, inf],
-                        [inf, 4.29865812, 4.29865812, inf],
-                        [inf, 4.29865812, 4.29865812, inf],
-                        [inf, inf, inf, inf],
+                        [nan, nan, nan, nan],
+                        [nan, 4.29865812, 4.29865812, nan],
+                        [nan, 4.29865812, 4.29865812, nan],
[…]
