### 2023-07-26 ce061f6
Adding checks to ensure NaN propagation
