### 2023-07-26 83af4ea
Add wcs offset methods to Observation
