### 2023-07-26 e281c31
Add tests for wcs offsets
