### 2023-07-26 f094a6e
Run checks.yml for all events
