+++ planetmapper/common.py
@@ -1,4 +1,4 @@
-__version__ = '1.7.7'
+__version__ = '1.8.0'
 __author__ = 'Oliver King'
 __url__ = 'https://github.com/ortk95/planetmapper'
 __license__ = 'MIT'
