### 2023-08-01 #5728 @editorialbot
[REVIEW]: PlanetMapper: A Python package for visualising, navigating and mapping Solar System observations
**Submitting author:** @ortk95 (<a href="http://orcid.org/0000-0002-6271-0062">Oliver King</a>)
**Repository:** https://github.com/ortk95/planetmapper
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.8.1
