### 2023-08-22 ortk95/planetmapper#271 @tedjohnson12
Data for GUI Examples
Hi, I am creating this issue as part of the JOSS review process.
I really like the `planetmapper` page on readthedocs, but I think it could benefit from a sample dataset that new users could download and use with the GUI. I see all of the API examples can be created using the code in `examples/`, but perhaps you could add an `examples/data` directory with the images of Europa from Earth (or a toy version if those data are proprietary or extremely large). I think that could provide enormous benefit to new users. Thanks!
