### 2023-08-22 #5728 @tedjohnson12
@ortk95 please see issue https://github.com/ortk95/planetmapper/issues/271
It is a suggestion about including some example data for new GUI users, but is not critical. Thanks!
