### 2023-08-23 ortk95/planetmapper#271 @ortk95
While doing this, it might be worth explicitly listing the full set of kernels needed to run the GUI examples somewhere in the documentation to help make stuff easier for new users
