### 2023-08-23 ef274ee
Adding example Europa dataset for use with GUI
Initial work for #271
