### 2023-08-23 ortk95/planetmapper#271 @ortk95
Added Europa FITS file in ef274ee. Still need to

- [ ] Add a bit more useful info to README - e.g. link back to readthedocs documentation
- [ ] Add link from docs to the example data
