### 2023-08-23 1130b26
Fixing type checks for latest pyright
