### 2023-08-24 1cd7190
Document example GUI Europa data file
#271
