### 2023-08-24 4212c1f
Update .readthedocs.yaml
Update to latest ubuntu and python versions
