### 2023-08-24 ortk95/planetmapper#271 @ortk95
Added example dataset in https://github.com/ortk95/planetmapper/tree/main/examples/gui_data and link to this dataset in https://planetmapper.readthedocs.io/en/latest/user_interface.html#fitting-an-observation
