### 2023-09-18 13c663b
Raise TypeError for `Observation.set_img_size`
#276
