### 2023-09-18 9b2f545
Add tests for `Observation.set_img_size`
