### 2023-09-18 f65260a
Prevent `nx`, `ny` or `sz` kwargs when creating `Observation` objects
