+++ tests/test_observation.py
@@ -24,17 +24,21 @@ class TestObservation(unittest.TestCase):
     def test_init(self) -> None:
         with self.assertRaises(ValueError):
             Observation()
-
         with self.assertRaises(ValueError):
             Observation('some/path', data=np.ones((5, 5)))
-
         with self.assertRaises(ValueError):
             Observation('some/path', header=fits.Header({'key': 'value'}))
-
         with self.assertRaises(ValueError):
             Observation(
                 'some/path', data=np.ones((5, 5)), header=fits.Header({'key': 'value'})
             )
+
+        with self.assertRaises(TypeError):
+            Observation(self.path, nx=1)
+        with self.assertRaises(TypeError):
+            Observation(self.path, ny=1)
+        with self.assertRaises(TypeError):
+            Observation(self.path, sz=1)
 
         with self.subTest('image.png+target+observer+utc'):
             path = os.path.join(common_testing.DATA_PATH, 'inputs', 'image.png')
