### 2023-09-18 9f228d5
Fix copying of Observation objects
