+++ planetmapper/body_xy.py
@@ -315,7 +315,11 @@ class BodyXY(Body):
 
     def _copy_options_to_other(self, other: Self) -> None:
         super()._copy_options_to_other(other)
-        other.set_img_size(*self.get_img_size())
+        if other.get_img_size() != self.get_img_size():
+            # Only call set_img_size if sizes are different so that Observation
+            # objects can be copied safely (as calling set_img_size on an Observation
+            # will raise a TypeError).
+            other.set_img_size(*self.get_img_size())
         other.set_disc_params(*self.get_disc_params())
         other.set_disc_method(self.get_disc_method())
 
@@ -670,7 +674,7 @@ class BodyXY(Body):
             ny: If specified, set the number of pixels in the y dimension.
 
         Raises:
-            TypeError if `set_img_size` is called on an :class:`Observation` instance.
+            TypeError: if `set_img_size` is called on an :class:`Observation` instance.
         """
         if nx is not None:
             self._nx = nx
