### 2023-09-18 ccb4376
Remove redundant code when copying BodyXY
`nx` and `ny` are already set by `_get_kwargs` when copying a `BodyXY` instance, so don't need to do `set_img_size` as part of `_copy_options_to_other`. See also https://coveralls.io/builds/62720797/source?filename=planetmapper%2Fbody_xy.py#L322
