### 2023-09-26 1977a90
Remove cloning of python-type-stubs
#234
