+++ pyproject.toml
@@ -4,12 +4,10 @@ exclude = ""
 # Manually set exclude="" to prevent black using .gitignore due
 # to black not parsing trailing slashes in .gitignore correctly
 # https://github.com/psf/black/issues/3694
-extend_exclude = "python-type-stubs"
 
 
 [tool.isort]
 profile = "black"
-extend_skip = ["python-type-stubs"]
 
 
 [tool.coverage.run]
@@ -20,9 +18,9 @@ omit = ["planetmapper/gui.py"]
 [tool.coverage.report]
 exclude_lines = [
     "pragma: no cover",
-    "^\\s*\\.\\.\\.$", # "..." lines (typing overloads etc.)
+    "^\\s*\\.\\.\\.$",     # "..." lines (typing overloads etc.)
     "^if TYPE_CHECKING:$",
-    ]
+]
 
 
 [tool.pylint.'MESSAGES CONTROL']
@@ -41,18 +39,15 @@ disable = [
     "too-few-public-methods",
     "too-many-ancestors",
     "too-many-return-statements",
-    ]
-enable = [
-    "useless-suppression",
-    ]
+]
+enable = ["useless-suppression"]
 
 [tool.pylint.miscellaneous]
-notes=["FIXME", "XXX"] # Allow TODO comments, but not FIXME or XXX
+notes = ["FIXME", "XXX"] # Allow TODO comments, but not FIXME or XXX
 
 
 [tool.pyright]
 include = ["planetmapper/*.py", "setup.py"]
-stubPath = "python-type-stubs/stubs"
 reportPropertyTypeMismatch = true
 reportConstantRedefinition = true
 reportIncompatibleMethodOverride = true
