### 2023-09-26 4ad05f9
Remove stubs from github workflow
