### 2023-10-02 7e41bbb
Update type hints
