### 2023-10-02 270eb6c
Add Python 3.12 to CI checks matrix
