+++ .github/workflows/checks.yml
@@ -47,6 +47,7 @@ jobs:
       uses: actions/setup-python@v4
       with:
         python-version: ${{ matrix.python-version }}
+        allow-prereleases: true
         cache: "pip"
     - name: Install dependencies
       run: |
@@ -69,6 +70,7 @@ jobs:
       uses: actions/setup-python@v4
       with:
         python-version: ${{ matrix.python-version }}
+        allow-prereleases: true
         cache: "pip"
     - name: Install dependencies
       run: |
