### 2023-10-04 9a82dc0
Add setuptools to workflow requirements
