+++ .github/workflows/checks.yml
@@ -52,7 +52,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install pylint
+        pip install pylint setuptools
         pip install -r requirements.txt
     - name: Analysing the code with pylint
       run: |
