### 2023-10-05 ortk95/planetmapper#280 @steo85it
Docu suggestions
The docs look great, very thorough and complete. I just have several minor suggestions:
- [x] In the wire-frame plot [example](https://planetmapper.readthedocs.io/en/latest/general_python_api.html), include the necessary kernels (de430 and leapseconds are not enough for even basic ones, e.g.,  `body = planetmapper.Body('saturn', '2020-01-01')`, probably because of its satellites - Saturn != Saturn Barycenter). As an alternative, use Mercury or Venus for such simple examples.
- [x] all examples regarding the attached europa.fits indicate europa.fits.gz (minor, still works with the uncompressed file)
- [x] include all necessary imports (planetmapper, pyplot) in the examples to facilitate the "copy/paste" by beginner users

[joss](https://github.com/openjournals/joss-reviews/issues/5728)
