### 2023-10-05 ortk95/planetmapper#281 @steo85it
JOSS paper review
The accompanying paper looks good to me: well explained motivation, reference to past/similar efforts, reference to SPICE as background tool, use cases... I only have a few suggestions:
l.7: an open source Python package that is used visualise, process .. --> an open source Python package to visualise, process ..
l.51: and is designed to be use.. --> and it is designed to be use..
l.56: and plan observing campaigns --> and plan observation campaigns
l.74: means that PlanetMapper can also be useful for --> also makes PlanetMapper well suited for

[joss](https://github.com/openjournals/joss-reviews/issues/5728)
