### 2023-10-05 ortk95/planetmapper#282 @steo85it
Code and testing
All examples and tests seem to run ok, installation works well both via pip or git clone.
However:
- [x] `./run_checks.sh` within the cloned dir results in a `black command not found` error. Am I doing something wrong?
- [x] `examples/make_logos.py` does not work with only the basic recommended kernels, see #280  (it should at least be documented which ones are needed)

[joss](https://github.com/openjournals/joss-reviews/issues/5728)
