### 2023-10-05 #5728 @steo85it
Hi @ortk95 ! Thanks for the very well developed piece of code and its manual, and sorry again for the really long delay!
I finally managed to go through paper, code, examples, etc... and found most of it to be ready for publication. I linked a few comments as "issues" on the planetmapper github (see above). Nothing critical, but possibly good to check/address.
