+++ joss/paper.md
@@ -20,7 +20,7 @@ bibliography: paper.bib
 
 # Summary
 <!-- A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience. -->
-`PlanetMapper` is an open source Python package that is used visualise, process and understand astronomical observations of Solar System objects, such as planets, moons and rings. Astronomers can use `PlanetMapper` to 'navigate' observations by calculating coordinate values (such as latitude and longitude) for each pixel in an observed image, and can map observations by projecting the observed data onto a map of the target body. Calculated values are exportable and directly accessible through a well documented API, allowing `PlanetMapper` to be used for custom analysis and processing. `PlanetMapper` can also be used to help generate publication quality figures, and has a Graphical User Interface to significantly simplify the processing of astronomical data. `PlanetMapper` can be applied to a wide range of datasets, including both amateur and professional ground-based observations, and data from space telescopes like Hubble and JWST.
+`PlanetMapper` is an open source Python package to visualise, process and understand astronomical observations of Solar System objects, such as planets, moons and rings. Astronomers can use `PlanetMapper` to 'navigate' observations by calculating coordinate values (such as latitude and longitude) for each pixel in an observed image, and can map observations by projecting the observed data onto a map of the target body. Calculated values are exportable and directly accessible through a well documented API, allowing `PlanetMapper` to be used for custom analysis and processing. `PlanetMapper` can also be used to help generate publication quality figures, and has a Graphical User Interface to significantly simplify the processing of astronomical data. `PlanetMapper` can be applied to a wide range of datasets, including both amateur and professional ground-based observations, and data from space telescopes like Hubble and JWST.
 
 # Statement of need
 <!-- A Statement of need section that clearly illustrates the research purpose of the software and places it in the context of related work. -->
@@ -36,17 +36,17 @@ Without answering these kinds of questions, it is often challenging to accuratel
 
 The NAIF SPICE Toolkit [@Acton2018] was developed by NASA to provide a standardised set of 'SPICE kernels', datasets containing the positions of Solar System objects, and a set of tools to interface with these kernels. This toolkit provides low level functions which can be combined to solve the problem of calculating the appearance of a target body. `PlanetMapper` is designed to significantly simplify the use of SPICE for planetary astronomers, effectively providing a high level interface to the toolkit. For example, the conversion between right ascension/declination coordinates (in the sky of the observer) to latitude/longitude coordinates (on the target body) requires calling ~10 SPICE functions, but can be done in a single function call with `PlanetMapper`. `PlanetMapper` makes use of the `SpiceyPy` package [@Annex2020] which provides a Python interface to the low level SPICE toolkit functions.
 
-Planetary astronomers have developed toolkits for image navigation, such as the `IDL` language based `DRM` [@Fletcher2009] and the `WinJUPOS` windows application for analysing Jupiter observations ([jupos.privat.t-online.de](http://jupos.privat.t-online.de/index.htm)). The USGS `ISIS` software [@ISIS] also provides a comprehensive set of tools for processing and navigating data from specific instruments on some spacecraft missions, but does not support generalised datasets or ground-based observations. `PlanetMapper` is the first general purpose Python package for navigating and mapping astronomical observations, and is designed to be used with any form of imaging data observing any Solar System object which has a SPICE kernel available.
[…]
