### 2023-10-05 ortk95/planetmapper#281 @ortk95
Thanks for the suggestions! I've incorporated them all in a2c4a48
