### 2023-10-05 ortk95/planetmapper#282 @ortk95
> ./run_checks.sh within the cloned dir results in a black command not found error. Am I doing something wrong?

The developemnt dependencies were never actually documented anywhere. I've now added an explicit `dev_requirements.txt` file, so all the development dependencies can be installed with `pip install -r dev_requirements.txt`. Also added associated documentation to signpost this.

66ab80b228e9197599fa6e142c7ef6cfa12c5551:
