### 2023-10-05 969bd2a
Correcting path for eurpoa.fits
Path is now consistent with example Europa data file (i.e. no .gz). Also added a note pointing to the example data file for users to play with.

#280
