### 2023-10-05 ortk95/planetmapper#280 @ortk95
> all examples regarding the attached europa.fits indicate europa.fits.gz (minor, still works with the uncompressed file)

Updated paths in 969bd2aa4aa679854a7a2094dfefbde002156ece
