### 2023-10-05 48c99c2
Update figure generation script to point to new example file
