### 2023-10-05 ortk95/planetmapper#280 @ortk95
>  include all necessary imports (planetmapper, pyplot) in the examples to facilitate the "copy/paste" by beginner users

Added imports in a2bb51bac181846eed05d5512c530b8c4fa602ad
