### 2023-10-05 cf1f338
Simplified paths in examples
No real need for the complex directory structures
