### 2023-10-05 fb36068
Document SPICE kernels required for examples
#280
