### 2023-10-05 ortk95/planetmapper#280 @ortk95
>In the wire-frame plot [example](https://planetmapper.readthedocs.io/en/latest/general_python_api.html), include the necessary kernels (de430 and leapseconds are not enough for even basic ones, e.g., body = planetmapper.Body('saturn', '2020-01-01'), probably because of its satellites - Saturn != Saturn Barycenter). As an alternative, use Mercury or Venus for such simple examples.

Documented the kernels needed to run the examples, and added a code snippet to download the required kernels in fb3606862f3227a64530ffe7fc110f96d68cdd10
