### 2023-10-05 ortk95/planetmapper#280 @ortk95
https://planetmapper.readthedocs.io/en/dev/general_python_api.html now documents the required kernels, includes imports for all code snippets and uses consistent filenames with the example data file
