### 2023-10-05 110719c
Add required kernels to make_logos.py
Include function to download the kernels needed to produce the logos

#282
