### 2023-10-05 ortk95/planetmapper#282 @ortk95
> examples/make_logos.py does not work with only the basic recommended kernels, see https://github.com/ortk95/planetmapper/issues/280 (it should at least be documented which ones are needed)

Added required kernels (& script to download them) in 110719ce0815c1f1a1ca82bf468afda8f243b1b1
