+++ examples/general_python_api.py
@@ -3,6 +3,8 @@
 """
 Script used to generate example plots on 
 https://planetmapper.readthedocs.io/en/latest/general_python_api.html
+
+Run download_spice_kernels() to download the kernels required for these examples.
 """
 import os
 import sys
