### 2023-10-05 #5728 @mbobra
@ortk95 Could you please [modify this Zenodo deposit](https://zenodo.org/record/8410701) so that that the author list and title on the deposit matches the author list and title of the JOSS paper (or, of course, feel free to cut a new release as well)?
