### 2023-10-09 fd598dc
Refactor projection string logic
Create generic function to produce projection strings, and correctly handle the strings for different +ve longitude directions.
