### 2023-10-09 2099b09
Add unit test for create_proj_string
