### 2023-10-09 1905839
Improve & correct generate_map_coordinates tests
