+++ tests/test_body_xy.py
@@ -828,7 +828,7 @@ class TestBodyXY(unittest.TestCase):
                 else:
                     self.assertEqual(a, b)
 
-        args: list[
+        args_limit: list[
             tuple[
                 tuple[float, float] | None,
                 tuple[float, float] | None,
@@ -867,7 +867,7 @@ class TestBodyXY(unittest.TestCase):
                 array([[-45.0, -45.0], [45.0, 45.0]]),
             ),
         ]
-        for xlim, ylim, lons_expected, lats_expected in args:
+        for xlim, ylim, lons_expected, lats_expected in args_limit:
             with self.subTest(xlim=xlim, ylim=ylim):
                 (
                     lons,
@@ -883,11 +883,11 @@ class TestBodyXY(unittest.TestCase):
                     np.array_equal(lons, lons_expected),
                     msg=f'{lons} <> {lons_expected}',
                 )
-                self.assertTrue(np.array_equal(lats, lats_expected))
-                self.assertTrue(np.array_equal(xx, lons_expected))
-                self.assertTrue(np.array_equal(yy, lats_expected))
-                self.assertEqual(info['xlim'], xlim)
-                self.assertEqual(info['ylim'], ylim)
+                self.assertTrue(np.array_equal(lats, lats_expected), msg=repr(lats))
+                self.assertTrue(np.array_equal(xx, lons_expected), msg=repr(xx))
+                self.assertTrue(np.array_equal(yy, lats_expected), msg=repr(yy))
+                self.assertEqual(info['xlim'], xlim, msg=repr(info))
+                self.assertEqual(info['ylim'], ylim, msg=repr(info))
 
         # Test reference
         args: list[tuple[MapKwargs, np.ndarray, np.ndarray, np.ndarray, np.ndarray]] = [
@@ -899,26 +899,90 @@ class TestBodyXY(unittest.TestCase):
                 array([[-28.5, -28.5, -28.5]]),
             ),
             (
-                MapKwargs(projection='orthographic', size=3),
-                array([[nan, nan, nan], [nan, 0.0, nan], [nan, nan, nan]]),
-                array([[nan, nan, nan], [nan, 0.0, nan], [nan, nan, nan]]),
-                array([[-1.01, 0.0, 1.01], [-1.01, 0.0, 1.01], [-1.01, 0.0, 1.01]]),
-                array([[-1.01, -1.01, -1.01], [0.0, 0.0, 0.0], [1.01, 1.01, 1.01]]),
+                MapKwargs(projection='orthographic', size=5),
+                array(
+                    [
+                        [nan, nan, nan, nan, nan],
+                        [nan, 36.87110893, 0.0, -36.87110893, nan],
+                        [nan, 30.33135236, 0.0, -30.33135236, nan],
+                        [nan, 36.87110893, 0.0, -36.87110893, nan],
+                        [nan, nan, nan, nan, nan],
+                    ]
+                ),
+                array(
+                    [
+                        [nan, nan, nan, nan, nan],
+                        [nan, -34.45624462, -34.45624462, -34.45624462, nan],
+                        [nan, 0.0, 0.0, 0.0, nan],
+                        [nan, 34.45624462, 34.45624462, 34.45624462, nan],
+                        [nan, nan, nan, nan, nan],
+                    ]
+                ),
+                array(
+                    [
+                        [-1.01, -0.505, 0.0, 0.505, 1.01],
+                        [-1.01, -0.505, 0.0, 0.505, 1.01],
+                        [-1.01, -0.505, 0.0, 0.505, 1.01],
+                        [-1.01, -0.505, 0.0, 0.505, 1.01],
+                        [-1.01, -0.505, 0.0, 0.505, 1.01],
+                    ]
+                ),
+                array(
+                    [
[…]
