### 2023-10-09 dad8211
Revert "Fix bug where pyproj transformations returned by `generate_map_coordinates` could have an inconsistent positive longitude direction with other parts of PlanetMapper"
