### 2023-10-10 #5728 @mbobra
> > btw @mbobra , can/should we add our orcid as reviewers, and if yes, how? Thanks!
>
> @steo85it That is a good question -- I don't know the answer. I'll investigate and get back to you. I'll start the paper acceptance process, but I won't finish the process before getting back to you.

@steo85it It looks like there is no process for reviewers to add their ORCID. I think you brought up an excellent point and I encourage you to open an issue at [openjournals/joss](https://github.com/openjournals/joss/issues). I'm really sorry we don't have that capability yet -- I absolutely agree that you should get as much credit as possible, especially since you're volunteering your time.
