### 2023-10-10 #5728 @mbobra
Thank you both so very much for your thorough reviews, @steo85it and @tedjohnson12! Your volunteer time keeps JOSS running and the JOSS team appreciates it very much ☀️
