### 2023-10-10 #5728 @ortk95
The proof looks good to me

Thanks very much for your reviews and helpful suggestions @tedjohnson12 & @steo85it, and thanks @mbobra for organising everything!
