### 2023-10-10 ortk95/planetmapper#295 @ortk95
Add links to JOSS paper
Once the JOSS paper is accepted & published, add relevant links to it:

- [x] Readme (badge & explicit section)
- [x] Main documentation pages
- [x] Source code (__init__ docstring & maybe any module level dunder constants)
- [x] CITATION file
- [x] setup.py urls
- [x] ? Can maybe add link to .zenodo.json

https://github.com/openjournals/joss-reviews/issues/5728
